fix s6-overlay shutdown

+++ web/action.py
@@ -326,7 +326,10 @@ class WebAction:
             os.system(
                 "ps -ef | grep -v grep | grep 'python run.py'|awk '{print $2}'|xargs kill -9")
         else:
-            os.system("pkill -f 'python3 run.py'")
+            if SystemUtils.check_process('pm2-runtime'):
+                os.system("pm2 restart NAStool")
+            else:
+                os.system("pkill -f 'python3 run.py'")
 
     @staticmethod
     def handle_message_job(msg, in_from=SearchType.OT, user_id=None, user_name=None):
